Extracted 'copy_velocities' to 'utils' module
